Refactor package prefix to solutions.fairdata
