+++ pom.xml
@@ -26,7 +26,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.fair.openrefine</groupId>
+    <groupId>solutions.fairdata.openrefine</groupId>
     <artifactId>metadata</artifactId>
     <version>0.1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
