+++ src/main/java/solutions/fairdata/openrefine/metadata/example/SampleUtil.java
@@ -20,18 +20,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.fair.openrefine.metadata.example;
+package solutions.fairdata.openrefine.metadata.example;
 
-import org.junit.jupiter.api.Test;
+public class SampleUtil {
+    // Just an example file
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-
-public class SampleUtilTest {
-    // Just an example test file
-
-    @Test
-    public void testSampleMethod() {
-        assertEquals(42, SampleUtil.sampleMethod());
+    static public int sampleMethod() {
+        return 42;
     }
 }
